replaced topicSearchRequest, moved functionality into backend
as a first step towards backend usage, topicSearchRequest was
used to directly query the backend for the very first request
concerning topics matching a simple search query 'query'.

The response by the backend is as defined in src/types/app.ts
to avoid further transformation steps in the webapp. Therefore
also src/state/dataAPI.ts had to be changed to take the results
more directly.